Emitir Certificado - Modificações
Emitir certificados está de acordo com os diagramas de 1 a 3 dos diagramas de colaboração de emitir certificado.

+++ src/main/java/br/unesp/fct/evcomp/repository/PresencaRepository.java
@@ -12,7 +12,7 @@ public interface PresencaRepository extends JpaRepository<RegistroDePresenca, In
     Optional<RegistroDePresenca> buscarPresencaPorAtividade(@Param("atividadeId") Integer atividadeId, @Param("participanteId") Integer participanteId);
 
     @Query("SELECT COUNT(p) FROM RegistroDePresenca p JOIN p.atividade a WHERE p.participante.id = :participanteId AND a.evento.id = :eventoId AND p.presente = true")
-    long contarPresencasNoEvento(@Param("participanteId") Integer participanteId, @Param("eventoId") Integer eventoId);
+    int contarPresencasNoEvento(@Param("participanteId") Integer participanteId, @Param("eventoId") Integer eventoId);
 
     @Query("SELECT p.atividade.id FROM RegistroDePresenca p WHERE p.participante.id = :participanteId AND p.presente = true")
     java.util.List<Integer> buscarAtividadesComPresencaPorParticipante(@Param("participanteId") Integer participanteId);
